pos: clear the warnings the build has been printing all along
Four warnings sat in every compile of the staking sources:

  minter.cpp:199  unused variable 'nNextSearch'  -- computed, then an
                  unconditional continue; the delay it implies was never
                  written, so the line goes
  multiwallet.cpp:63  signed/unsigned comparison in the wallet-matching
                  loop
  stake.cpp:318   'scriptPubKeyKernel' shadows a previous local; the
                  outer declaration was never used

and one silent wrong answer: the "Wallet %d" lines in the staking loop
printed a size_t that was initialised to 0 and never incremented, so
every locked-wallet and low-balance message named wallet 0 regardless of
which of the eight it was.

Three warnings in these files are deliberately left: nFees in
CreateCoinStake, and nTime/nHeight in SelectCoinsForStaking. Each marks
a real unfixed defect -- the stake reward drops the fees it is handed,
and coin selection ignores the time it is given in favour of the wall
clock. Silencing them would hide the bugs.

+++ src/pos/multiwallet.cpp
@@ -60,7 +60,7 @@ void MultiwalletMaintenance()
         CStakeWallet wallet(p.get(), (Consensus::Params&) params);
         wallet.StakingDisabled();
         //use temporary copy of old wallet list to determine which wallets were staking
-        for (int i=0; i<temp_wallets.size(); i++) {
+        for (size_t i = 0; i < temp_wallets.size(); i++) {
             if (wallet.GetName() == temp_wallets[i].GetName()) {
                 if (temp_wallets[i].CanStake()) {
                     wallet.StakingEnabled();
